Added `deserializeReceipt`
